[MSE] Don't use fudgeFactor when computing hasCurrentTime()
This change solves the following seek issue: During a seek to 2511.0, data is
pushed to the pipeline. The last sample for audiotrack had PTS=2510.913086.
This caused MediaSource::hasCurrentTime() to return true (it considers
fudgeFactor), but 2511.0 isn't actually buffered.

MediaSource::monitorSourceBuffers thinks that currentTime is available and calls
completeSeek(), which in turn silently fails to sourceBuffer->seekToTime(m_pendingSeekTime).
m_pendingSeekTime is unconditionally set to MediaTime::invalidTime() and the
further seek fails. No more data will be pushed to the playback pipeline because
it's "seeking" and completeSeek() is only called from monitorSourceBuffers()
only when m_pendingSeekTime.isValid().

This change avoids using the fudgeFactor when computing hasCurrentTime().